+++ docs/.vitepress/locale.ts
@@ -55,6 +55,11 @@ export function getLocaleConfig(lang: string) {
       activeMatch: 'features',
     },
     {
+      text: 'Volar',
+      link: `${urlPrefix}/volar/template-ref`,
+      activeMatch: 'volar',
+    },
+    {
       text: t('Interactive Example'),
       link: `${urlPrefix}/interactive/`,
       activeMatch: 'interactive',
@@ -227,6 +232,24 @@ export function getLocaleConfig(lang: string) {
             {
               text: 'booleanProp',
               link: `${urlPrefix}/features/boolean-prop`,
+            },
+            {
+              text: 'scriptLang',
+              link: `${urlPrefix}/features/script-lang`,
+            },
+          ],
+        },
+      ],
+    },
+    {
+      text: 'Volar',
+      items: [
+        {
+          text: t('Experimental'),
+          items: [
+            {
+              text: 'templateRef',
+              link: `${urlPrefix}/volar/template-ref`,
             },
           ],
         },