+++ playground/vue2/src/examples/script-lang/comp.vue
@@ -1,0 +1,11 @@
+<script setup>
+defineProps<{
+  foo: string
+}>()
+</script>
+
+<template>
+  <div>
+    {{ foo }}
+  </div>
+</template>