+++ playground/vue3/src/examples/script-lang/index.vue
@@ -1,0 +1,7 @@
+<script setup>
+import Comp from './comp.vue'
+</script>
+
+<template>
+  <Comp foo="foo" />
+</template>